Bugfix
If the user has deleted the user owen age or like that. There will be a error about argument 1 should be a int and null is given. Therefor it will in that case use 0 in all three. The age will be the current age but it is much better end error.